refactor(1932): refactor/adjustments after code review
