+++ src/plugins/account/services/account-cleanup.service.interface.ts
@@ -1,12 +1,8 @@
 import type { SupportedNetwork } from '@/core/types/shared.types';
-
-export interface AccountDeleteLocalStateInput {
-  accountId: string;
-}
-
-export interface AccountDeleteKmsCleanupInput {
-  keyRefId: string;
-}
+import type {
+  AccountDeleteKmsCleanupInput,
+  AccountDeleteLocalStateInput,
+} from '@/plugins/account/services/account-cleanup.types';
 
 export interface AccountCleanupService {
   removeAccountFromLocalState(
