+++ src/plugins/account/services/account-cleanup.types.ts
@@ -1,0 +1,7 @@
+export interface AccountDeleteLocalStateInput {
+  accountId: string;
+}
+
+export interface AccountDeleteKmsCleanupInput {
+  keyRefId: string;
+}
