+++ src/plugins/eip712/services/eip712-verification.service.interface.ts
@@ -1,24 +1,8 @@
-import type {
-  Credential,
-  KeyManager,
-} from '@/core/services/kms/kms-types.interface';
-import type { SupportedNetwork } from '@/core/types/shared.types';
 import type { Eip712VerifyOutput } from '@/plugins/eip712/commands/verify/output';
-import type { ExpectedSignerType } from '@/plugins/eip712/commands/verify/types';
-
-export interface VerifyEcdsaParams {
-  signature: string;
-  hash: string;
-  expectedSigner: ExpectedSignerType | undefined;
-  network: SupportedNetwork;
-}
-
-export interface VerifyEd25519Params {
-  keyManager: KeyManager;
-  credential: Credential | undefined;
-  signature: string;
-  hash: string;
-}
+import type {
+  VerifyEcdsaParams,
+  VerifyEd25519Params,
+} from '@/plugins/eip712/services/eip712-verification.types';
 
 export interface Eip712VerificationService {
   verifyEcdsa(params: VerifyEcdsaParams): Promise<Eip712VerifyOutput>;
