+++ src/plugins/hbar/services/hbar-allowance-query.types.ts
@@ -1,0 +1,6 @@
+import type { HbarAllowanceInfo } from '@/core/services/mirrornode/types';
+
+export interface HbarAllowanceFetchResult {
+  allowances: HbarAllowanceInfo[];
+  hasMore: boolean;
+}
