+++ src/plugins/schedule/services/schedule-sync.service.interface.ts
@@ -1,6 +1,6 @@
+import type { ScheduledTransactionData } from '@/core/schemas/common-schemas';
 import type { KeyManager } from '@/core/services/kms/kms-types.interface';
 import type { SupportedNetwork } from '@/core/types/shared.types';
-import type { ScheduledTransactionData } from '@/plugins/schedule/schema';
 
 export interface ScheduleSyncService {
   markExecutionStatusFromMirror(
