--- src/plugins/token/hooks/token-associate-state/index.ts
@@ -1,1 +1,0 @@
-export { TokenAssociateStateHook } from './handler';
