--- src/plugins/token/hooks/token-dissociate-state/index.ts
@@ -1,1 +1,0 @@
-export { TokenDissociateStateHook } from './handler';
