--- src/plugins/token/hooks/token-dissociate-state/types.ts
@@ -1,15 +1,0 @@
-import { z } from 'zod';
-
-import {
-  NetworkSchema,
-  ResolvedAccountCredentialSchema,
-} from '@/core/schemas/common-schemas';
-import { keyManagerNameSchema } from '@/core/services/kms/kms-types.interface';
-
-export const DissociateNormalizedParamsSchema = z.object({
-  network: NetworkSchema,
-  tokenId: z.string(),
-  account: ResolvedAccountCredentialSchema,
-  keyManager: keyManagerNameSchema,
-  keyRefIds: z.array(z.string()),
-});
