+++ src/plugins/token/services/token-pending-airdrops.service.interface.ts
@@ -1,24 +1,5 @@
 import type { SupportedNetwork } from '@/core/types/shared.types';
-
-export type PendingAirdropType = 'FUNGIBLE' | 'NFT';
-
-export interface PendingAirdropEntry {
-  tokenId: string;
-  tokenName: string;
-  tokenSymbol: string;
-  senderId: string;
-  type: PendingAirdropType;
-  amount?: number;
-  serialNumber?: number;
-}
-
-export interface TokenPendingAirdropsResult {
-  account: string;
-  network: SupportedNetwork;
-  airdrops: PendingAirdropEntry[];
-  hasMore: boolean;
-  total: number;
-}
+import type { TokenPendingAirdropsResult } from '@/plugins/token/services/token-pending-airdrops.types';
 
 export interface TokenPendingAirdropsService {
   getPendingAirdrops(
