+++ src/plugins/topic/services/topic-alias.types.ts
@@ -1,0 +1,8 @@
+import type { SupportedNetwork } from '@/core/types/shared.types';
+
+export interface RegisterTopicAliasParams {
+  alias: string;
+  network: SupportedNetwork;
+  topicId: string;
+  createdAt: string;
+}
